Replace long/unsigned long with fixed-width int32_t/uint32_t
Map Web IDL types to portable C++ fixed-width integers:
- Web IDL 'long' (signed 32-bit) -> int32_t
- Web IDL 'unsigned long' (unsigned 32-bit) -> uint32_t

Plain 'long' is 32-bit on Windows but 64-bit on Linux, making it
non-portable for cross-platform DOM interfaces.

Affected interfaces:
- SVGTextContentElement: all char index/count methods and syncTextLayout
- TextLayout/PangoTextLayout/SkiaTextLayout: GetCharOffsetX parameter
- KeyboardEvent: location constants and accessor
- UIEvent: detail field and accessor
- WheelEvent: deltaMode constants, field, and accessor
- Window: HandleKeyDown/HandleKeyUp/HandleWheel parameters
- SVGLengthList/SVGNumberList: length, numberOfItems, index parameters

Added explicit #include <cstdint> where needed.

+++ core/PangoTextLayout.cpp
@@ -165,7 +165,7 @@ namespace AeonGUI
         return static_cast<double> ( baseline ) / PANGO_SCALE;
     }
 
-    double PangoTextLayout::GetCharOffsetX ( long aIndex ) const
+    double PangoTextLayout::GetCharOffsetX ( int32_t aIndex ) const
     {
         PangoRectangle pos;
         pango_layout_index_to_pos ( mLayout, static_cast<int> ( aIndex ), &pos );
